Renamed components and layouts
