+++ my_travel_guide/lib/firebase/landmark.dart
@@ -1,5 +1,3 @@
-import 'package:flutter/material.dart';
-
 class Doodle {
   final String name;
   final String time;
