+++ my_travel_guide/lib/firebase/user_timeline_landmarks.dart
@@ -2,7 +2,7 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:my_travel_guide/authentication/google_sign_in.dart';
-import '../firebase/data.dart';
+import '../firebase/landmark.dart';
 
 
 final databaseReference = Firestore.instance;
